+++ app/routes/components/DatePicker/demos/Demo01.tsx
@@ -1,5 +1,15 @@
+import type dayjs from "dayjs";
+import { useState } from "react";
 import { DatePicker } from "@/components/DatePicker";
+import { Flex } from "@/components/Flex";
 
 export default function Demo() {
-  return <DatePicker />;
+  const [date, setDate] = useState<dayjs.Dayjs | null>(null);
+
+  return (
+    <Flex dir="column" gap="1rem" alignItems="center">
+      <DatePicker date={date} onDateChange={setDate} />
+      <span>{date?.format("YYYY-MM-DD HH:mm:ss")}</span>
+    </Flex>
+  );
 }