+++ src/components/DatePicker/index.ts
@@ -1,3 +1,3 @@
-// DatePicker from copy-ui @ 2025-10-25.
+// DatePicker from copy-ui @ 2025-10-28.
 
 export { DatePicker } from "./date-picker";