+++ src/components/DatePicker/CHANGELOG.md
@@ -1,1 +1,2 @@
 - 2025-10-25: Initial version.
+- 2025-11-28: Support `date` and `onDateChange` props to make the component controllable.